backend/requirements.txt
Django==5.1.5
djangorestframework==3.15.2
django-cors-headers==4.6.0
celery==5.4.0
redis==5.2.1
requests==2.32.3
cryptography==44.0.0
psycopg[binary]==3.2.13
python-dotenv==1.0.1
gunicorn==23.0.0
whitenoise==6.9.0
